scripts/data_pipeline/requirements.txt
# CardScope 数据管道依赖
# Python 3.9+

supabase>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
